+++ server/src/core/routines.interface.ts
@@ -1,5 +1,7 @@
 import { CompletionItem } from "vscode-languageserver";
 
+// all keys are strings of the index to access our item
+
 export interface IPropertyItem {
   [key: string]: string; // key name of property, value is the IDL data type
 }
@@ -16,6 +18,10 @@ export interface IProcedures {
   [key: string]: boolean; // key is string of data from above
 }
 
+export interface IMethods {
+  [key: string]: boolean; // key is string of data from above
+}
+
 export interface ILinks {
   [key: string]: string; // key is string of data from above, value is sub-link on docs center to find more information
 }
@@ -26,4 +32,5 @@ export interface IRoutines {
   links: ILinks;
   functions: IFunctions;
   procedures: IProcedures;
+  methods: IMethods;
 }