+++ server/src/core/search.interface.ts
@@ -1,0 +1,15 @@
+import { CompletionItem } from "vscode-languageserver";
+
+export interface IQuickLookup {
+  functions: CompletionItem[];
+  procedures: CompletionItem[];
+  functionMethods: CompletionItem[];
+  procedureMethods: CompletionItem[];
+}
+
+export interface IQuickSearchLookup {
+  functions: Fuzzysort.Prepared[];
+  procedures: Fuzzysort.Prepared[];
+  functionMethods: Fuzzysort.Prepared[];
+  procedureMethods: Fuzzysort.Prepared[];
+}